requirements.txt
fastapi
uvicorn
pydantic
httpx
pandas
numpy
pyjwt
mcp>=1.6.0
pydantic-ai

python-jose 
passlib 
python-multipart 
pyjwt 
python-dotenv>=0.19.0
langfuse
boto3>=1.26.0
google-cloud-storage>=2.13.0
google-auth>=2.27.0